+++ relay-general/src/store/mod.rs
@@ -61,6 +61,9 @@ pub struct StoreConfig {
 
     /// Emit additional span attributes based on given configuration.
     pub span_attributes: BTreeSet<SpanAttribute>,
+
+    /// The SDK's sample rate as communicated via envelope headers.
+    pub client_sample_rate: Option<f64>,
 }
 
 /// The processor that normalizes events for store.